+++ Core/src/main/java/com/plotsquared/core/PlotSquared.java
@@ -74,6 +74,7 @@ import com.plotsquared.core.util.LegacyConverter;
 import com.plotsquared.core.util.MainUtil;
 import com.plotsquared.core.util.MathMan;
 import com.plotsquared.core.util.ReflectionUtils;
+import com.plotsquared.core.util.RegionManager;
 import com.plotsquared.core.util.SchematicHandler;
 import com.plotsquared.core.util.SetupUtils;
 import com.plotsquared.core.util.StringMan;
@@ -273,6 +274,7 @@ public class PlotSquared {
             GlobalBlockQueue.IMP.runTask();
             // Set chunk
             ChunkManager.manager = this.IMP.initChunkManager();
+            RegionManager.manager = this.IMP.initRegionManager();
             // Schematic handler
             SchematicHandler.manager = this.IMP.initSchematicHandler();
             // Chat
