+++ Core/src/main/java/com/plotsquared/core/command/Backup.java
@@ -53,11 +53,11 @@ import java.util.stream.IntStream;
 import java.util.stream.Stream;
 
 @CommandDeclaration(command = "backup",
-usage = "/plot backup <save|list|load>",
-description = "Manage plot backups",
-category = CommandCategory.SETTINGS,
-requiredType = RequiredType.PLAYER,
-permission = "plots.backup")
+    usage = "/plot backup <save|list|load>",
+    description = "Manage plot backups",
+    category = CommandCategory.SETTINGS,
+    requiredType = RequiredType.PLAYER,
+    permission = "plots.backup")
 public final class Backup extends Command {
 
     public Backup() {
@@ -69,7 +69,8 @@ public final class Backup extends Command {
         return true;
     }
 
-    @Override public CompletableFuture<Boolean> execute(PlotPlayer player, String[] args,
+    @Override
+    public CompletableFuture<Boolean> execute(PlotPlayer player, String[] args,
         RunnableVal3<Command, Runnable, Runnable> confirm,
         RunnableVal2<Command, CommandResult> whenDone) throws CommandException {
         if (args.length == 0 || !Arrays.asList("save", "list", "load")
@@ -82,23 +83,28 @@ public final class Backup extends Command {
     @Override public Collection<Command> tab(PlotPlayer player, String[] args, boolean space) {
         if (args.length == 1) {
             return Stream.of("save", "list", "load")
-                          .filter(value -> value.startsWith(args[0].toLowerCase(Locale.ENGLISH)))
-                          .map(value -> new Command(null, false, value, "", RequiredType.NONE, null) {})
-                          .collect(Collectors.toList());
+                .filter(value -> value.startsWith(args[0].toLowerCase(Locale.ENGLISH)))
+                .map(value -> new Command(null, false, value, "", RequiredType.NONE, null) {
+                }).collect(Collectors.toList());
         } else if (args[0].equalsIgnoreCase("load")) {
 
             final Plot plot = player.getCurrentPlot();
             if (plot != null) {
-                final BackupProfile backupProfile = Objects.requireNonNull(PlotSquared.imp()).getBackupManager().getProfile(plot);
+                final BackupProfile backupProfile =
+                    Objects.requireNonNull(PlotSquared.imp()).getBackupManager().getProfile(plot);
                 if (backupProfile instanceof PlayerBackupProfile) {
-                    final CompletableFuture<List<com.plotsquared.core.backup.Backup>> backupList = backupProfile.listBackups();
+                    final CompletableFuture<List<com.plotsquared.core.backup.Backup>> backupList =
+                        backupProfile.listBackups();
                     if (backupList.isDone()) {
-                        final List<com.plotsquared.core.backup.Backup> backups = backupList.getNow(new ArrayList<>());
+                        final List<com.plotsquared.core.backup.Backup> backups =
+                            backupList.getNow(new ArrayList<>());
                         if (backups.isEmpty()) {
                             return new ArrayList<>();
                         }
-                        return IntStream.range(1, 1 + backups.size()).mapToObj(i -> new Command(null, false, Integer.toString(i),
-                            "", RequiredType.NONE, null) {}).collect(Collectors.toList());
+                        return IntStream.range(1, 1 + backups.size()).mapToObj(
+                            i -> new Command(null, false, Integer.toString(i), "",
+                                RequiredType.NONE, null) {
+                            }).collect(Collectors.toList());
 
                     }
                 }
@@ -108,11 +114,11 @@ public final class Backup extends Command {
     }
 
     @CommandDeclaration(command = "save",
-    usage = "/plot backup save",
-    description = "Create a plot backup",
-    category = CommandCategory.SETTINGS,
-    requiredType = RequiredType.PLAYER,
-    permission = "plots.backup.save")
+        usage = "/plot backup save",
+        description = "Create a plot backup",
+        category = CommandCategory.SETTINGS,
+        requiredType = RequiredType.PLAYER,
+        permission = "plots.backup.save")
     public void save(final Command command, final PlotPlayer player, final String[] args,
         final RunnableVal3<Command, Runnable, Runnable> confirm,
         final RunnableVal2<Command, CommandResult> whenDone) {
@@ -120,15 +126,20 @@ public final class Backup extends Command {
         if (plot == null) {
             sendMessage(player, Captions.NOT_IN_PLOT);
         } else if (!plot.hasOwner()) {
-            sendMessage(player, Captions.BACKUP_IMPOSSIBLE, Captions.GENERIC_UNOWNED.getTranslated());
+            sendMessage(player, Captions.BACKUP_IMPOSSIBLE,
+                Captions.GENERIC_UNOWNED.getTranslated());
         } else if (plot.isMerged()) {
-            sendMessage(player, Captions.BACKUP_IMPOSSIBLE, Captions.GENERIC_MERGED.getTranslated());
-        } else if (!plot.isOwner(player.getUUID()) && !Permissions.hasPermission(player, Captions.PERMISSION_ADMIN_BACKUP_OTHER)) {
+            sendMessage(player, Captions.BACKUP_IMPOSSIBLE,
+                Captions.GENERIC_MERGED.getTranslated());
+        } else if (!plot.isOwner(player.getUUID()) && !Permissions
+            .hasPermission(player, Captions.PERMISSION_ADMIN_BACKUP_OTHER)) {
             sendMessage(player, Captions.NO_PERMISSION, Captions.PERMISSION_ADMIN_BACKUP_OTHER);
         } else {
-            final BackupProfile backupProfile = Objects.requireNonNull(PlotSquared.imp()).getBackupManager().getProfile(plot);
+            final BackupProfile backupProfile =
+                Objects.requireNonNull(PlotSquared.imp()).getBackupManager().getProfile(plot);
             if (backupProfile instanceof NullBackupProfile) {
-                sendMessage(player, Captions.BACKUP_IMPOSSIBLE, Captions.GENERIC_OTHER.getTranslated());
+                sendMessage(player, Captions.BACKUP_IMPOSSIBLE,
+                    Captions.GENERIC_OTHER.getTranslated());
             } else {
                 backupProfile.createBackup().whenComplete((backup, throwable) -> {
                     if (throwable != null) {
@@ -151,40 +162,47 @@ public final class Backup extends Command {
     public void list(final Command command, final PlotPlayer player, final String[] args,
         final RunnableVal3<Command, Runnable, Runnable> confirm,
         final RunnableVal2<Command, CommandResult> whenDone) {
-            final Plot plot = player.getCurrentPlot();
-            if (plot == null) {
-                sendMessage(player, Captions.NOT_IN_PLOT);
-            } else if (!plot.hasOwner()) {
-                sendMessage(player, Captions.BACKUP_IMPOSSIBLE, Captions.GENERIC_UNOWNED.getTranslated());
-            } else if (plot.isMerged()) {
-                sendMessage(player, Captions.BACKUP_IMPOSSIBLE, Captions.GENERIC_MERGED.getTranslated());
-            } else if (!plot.isOwner(player.getUUID()) && !Permissions.hasPermission(player, Captions.PERMISSION_ADMIN_BACKUP_OTHER)) {
-                sendMessage(player, Captions.NO_PERMISSION, Captions.PERMISSION_ADMIN_BACKUP_OTHER);
+        final Plot plot = player.getCurrentPlot();
+        if (plot == null) {
+            sendMessage(player, Captions.NOT_IN_PLOT);
+        } else if (!plot.hasOwner()) {
+            sendMessage(player, Captions.BACKUP_IMPOSSIBLE,
+                Captions.GENERIC_UNOWNED.getTranslated());
+        } else if (plot.isMerged()) {
+            sendMessage(player, Captions.BACKUP_IMPOSSIBLE,
+                Captions.GENERIC_MERGED.getTranslated());
+        } else if (!plot.isOwner(player.getUUID()) && !Permissions
+            .hasPermission(player, Captions.PERMISSION_ADMIN_BACKUP_OTHER)) {
+            sendMessage(player, Captions.NO_PERMISSION, Captions.PERMISSION_ADMIN_BACKUP_OTHER);
+        } else {
+            final BackupProfile backupProfile =
+                Objects.requireNonNull(PlotSquared.imp()).getBackupManager().getProfile(plot);
+            if (backupProfile instanceof NullBackupProfile) {
+                sendMessage(player, Captions.BACKUP_IMPOSSIBLE,
+                    Captions.GENERIC_OTHER.getTranslated());
             } else {
-                final BackupProfile backupProfile = Objects.requireNonNull(PlotSquared.imp()).getBackupManager().getProfile(plot);
-                if (backupProfile instanceof NullBackupProfile) {
-                    sendMessage(player, Captions.BACKUP_IMPOSSIBLE, Captions.GENERIC_OTHER.getTranslated());
-                } else {
-                    backupProfile.listBackups().whenComplete((backups, throwable) -> {
-                        if (throwable != null) {
-                            sendMessage(player, Captions.BACKUP_LIST_FAILED, throwable.getMessage());
-                            throwable.printStackTrace();
-                        } else {
-                            sendMessage(player, Captions.BACKUP_LIST_HEADER, plot.getId().toCommaSeparatedString());
-                            try {
-                                for (int i = 0; i < backups.size(); i++) {
-                                    sendMessage(player, Captions.BACKUP_LIST_ENTRY, Integer.toString(i + 1),
-                                        DateTimeFormatter.RFC_1123_DATE_TIME.format(ZonedDateTime
-                                                .ofInstant(Instant.ofEpochMilli(backups.get(i).getCreationTime()), ZoneId
-                                                .systemDefault())));
-                                }
-                            } catch (final Exception e) {
-                                e.printStackTrace();
+                backupProfile.listBackups().whenComplete((backups, throwable) -> {
+                    if (throwable != null) {
+                        sendMessage(player, Captions.BACKUP_LIST_FAILED, throwable.getMessage());
+                        throwable.printStackTrace();
+                    } else {
+                        sendMessage(player, Captions.BACKUP_LIST_HEADER,
+                            plot.getId().toCommaSeparatedString());
+                        try {
+                            for (int i = 0; i < backups.size(); i++) {
+                                sendMessage(player, Captions.BACKUP_LIST_ENTRY,
+                                    Integer.toString(i + 1), DateTimeFormatter.RFC_1123_DATE_TIME
+                                        .format(ZonedDateTime.ofInstant(
+                                            Instant.ofEpochMilli(backups.get(i).getCreationTime()),
+                                            ZoneId.systemDefault())));
                             }
+                        } catch (final Exception e) {
+                            e.printStackTrace();
                         }
-                    });
-                }
+                    }
+                });
             }
+        }
     }
 
     @CommandDeclaration(command = "load",
@@ -200,11 +218,14 @@ public final class Backup extends Command {
         if (plot == null) {
             sendMessage(player, Captions.NOT_IN_PLOT);
         } else if (!plot.hasOwner()) {
-            sendMessage(player, Captions.BACKUP_IMPOSSIBLE, Captions.GENERIC_UNOWNED.getTranslated());
+            sendMessage(player, Captions.BACKUP_IMPOSSIBLE,
+                Captions.GENERIC_UNOWNED.getTranslated());
         } else if (plot.isMerged()) {
-            sendMessage(player, Captions.BACKUP_IMPOSSIBLE, Captions.GENERIC_MERGED.getTranslated());
+            sendMessage(player, Captions.BACKUP_IMPOSSIBLE,
+                Captions.GENERIC_MERGED.getTranslated());
             sendMessage(player, Captions.BACKUP_IMPOSSIBLE, "merged");
-        } else if (!plot.isOwner(player.getUUID()) && !Permissions.hasPermission(player, Captions.PERMISSION_ADMIN_BACKUP_OTHER)) {
+        } else if (!plot.isOwner(player.getUUID()) && !Permissions
+            .hasPermission(player, Captions.PERMISSION_ADMIN_BACKUP_OTHER)) {
             sendMessage(player, Captions.NO_PERMISSION, Captions.PERMISSION_ADMIN_BACKUP_OTHER);
         } else if (args.length == 0) {
             sendMessage(player, Captions.BACKUP_LOAD_USAGE);
@@ -216,9 +237,11 @@ public final class Backup extends Command {
                 sendMessage(player, Captions.NOT_A_NUMBER, args[0]);
                 return;
             }
-            final BackupProfile backupProfile = Objects.requireNonNull(PlotSquared.imp()).getBackupManager().getProfile(plot);
+            final BackupProfile backupProfile =
+                Objects.requireNonNull(PlotSquared.imp()).getBackupManager().getProfile(plot);
             if (backupProfile instanceof NullBackupProfile) {
-                sendMessage(player, Captions.BACKUP_IMPOSSIBLE, Captions.GENERIC_OTHER.getTranslated());
+                sendMessage(player, Captions.BACKUP_IMPOSSIBLE,
+                    Captions.GENERIC_OTHER.getTranslated());
             } else {
                 backupProfile.listBackups().whenComplete((backups, throwable) -> {
                     if (throwable != null) {
@@ -226,20 +249,26 @@ public final class Backup extends Command {
                         throwable.printStackTrace();
                     } else {
                         if (number < 1 || number > backups.size()) {
-                            sendMessage(player, Captions.BACKUP_LOAD_FAILURE, Captions.GENERIC_INVALID_CHOICE.getTranslated());
+                            sendMessage(player, Captions.BACKUP_LOAD_FAILURE,
+                                Captions.GENERIC_INVALID_CHOICE.getTranslated());
                         } else {
-                            final com.plotsquared.core.backup.Backup backup = backups.get(number - 1);
-                            if (backup == null || backup.getFile() == null || !Files.exists(backup.getFile())) {
-                                sendMessage(player, Captions.BACKUP_LOAD_FAILURE, Captions.GENERIC_INVALID_CHOICE.getTranslated());
+                            final com.plotsquared.core.backup.Backup backup =
+                                backups.get(number - 1);
+                            if (backup == null || backup.getFile() == null || !Files
+                                .exists(backup.getFile())) {
+                                sendMessage(player, Captions.BACKUP_LOAD_FAILURE,
+                                    Captions.GENERIC_INVALID_CHOICE.getTranslated());
                             } else {
-                                CmdConfirm.addPending(player, "/plot backup load " + number, () ->
-                                    backupProfile.restoreBackup(backup).whenComplete((n, error) -> {
-                                   if (error != null) {
-                                       sendMessage(player, Captions.BACKUP_LOAD_FAILURE, error.getMessage());
-                                   } else {
-                                       sendMessage(player, Captions.BACKUP_LOAD_SUCCESS);
-                                   }
-                                }));
+                                CmdConfirm.addPending(player, "/plot backup load " + number,
+                                    () -> backupProfile.restoreBackup(backup)
+                                        .whenComplete((n, error) -> {
+                                            if (error != null) {
+                                                sendMessage(player, Captions.BACKUP_LOAD_FAILURE,
+                                                    error.getMessage());
+                                            } else {
+                                                sendMessage(player, Captions.BACKUP_LOAD_SUCCESS);
+                                            }
+                                        }));
                             }
                         }
                     }
