+++ Core/src/main/java/com/plotsquared/core/command/Set.java
@@ -81,8 +81,8 @@ public class Set extends SubCommand {
                     StringMan.join(Arrays.copyOfRange(args, 1, args.length), ",").trim();
 
                 final List<String> forbiddenTypes = Settings.General.INVALID_BLOCKS;
-                if (!Permissions.hasPermission(player, Captions.PERMISSION_ADMIN_ALLOW_UNSAFE) &&
-                    !forbiddenTypes.isEmpty()) {
+                if (!Permissions.hasPermission(player, Captions.PERMISSION_ADMIN_ALLOW_UNSAFE)
+                    && !forbiddenTypes.isEmpty()) {
                     for (String forbiddenType : forbiddenTypes) {
                         forbiddenType = forbiddenType.toLowerCase(Locale.ENGLISH);
                         if (forbiddenType.startsWith("minecraft:")) {
@@ -96,11 +96,14 @@ public class Set extends SubCommand {
 
                             if (blockType.startsWith("##")) {
                                 try {
-                                    final BlockCategory category = BlockCategory.REGISTRY.get(blockType.substring(2).toLowerCase(Locale.ENGLISH));
-                                    if (category == null || !category.contains(BlockTypes.get(forbiddenType))) {
+                                    final BlockCategory category = BlockCategory.REGISTRY
+                                        .get(blockType.substring(2).toLowerCase(Locale.ENGLISH));
+                                    if (category == null || !category
+                                        .contains(BlockTypes.get(forbiddenType))) {
                                         continue;
                                     }
-                                } catch (final Throwable ignored) {}
+                                } catch (final Throwable ignored) {
+                                }
                             } else if (!blockType.contains(forbiddenType)) {
                                 continue;
                             }
@@ -147,13 +150,13 @@ public class Set extends SubCommand {
             }
 
             @Override
-            public Collection<Command> tab(final PlotPlayer player, final String[] args, final boolean space) {
-                return PatternUtil.getSuggestions(player,  StringMan.join(args, ",").trim())
-                    .stream()
+            public Collection<Command> tab(final PlotPlayer player, final String[] args,
+                final boolean space) {
+                return PatternUtil.getSuggestions(player, StringMan.join(args, ",").trim()).stream()
                     .map(value -> value.toLowerCase(Locale.ENGLISH).replace("minecraft:", ""))
                     .filter(value -> value.startsWith(args[0].toLowerCase(Locale.ENGLISH)))
-                    .map(value -> new Command(null, false, value, "", RequiredType.NONE, null) {})
-                    .collect(Collectors.toList());
+                    .map(value -> new Command(null, false, value, "", RequiredType.NONE, null) {
+                    }).collect(Collectors.toList());
             }
         };
     }
@@ -197,13 +200,16 @@ public class Set extends SubCommand {
         return noArgs(player);
     }
 
-    @Override public Collection<Command> tab(final PlotPlayer player, final String[] args, final boolean space) {
+    @Override
+    public Collection<Command> tab(final PlotPlayer player, final String[] args,
+        final boolean space) {
         if (args.length == 1) {
             return Stream
-                .of("biome", "alias", "home", "main", "floor", "air", "all", "border", "wall", "outline", "middle")
+                .of("biome", "alias", "home", "main", "floor", "air", "all", "border", "wall",
+                    "outline", "middle")
                 .filter(value -> value.startsWith(args[0].toLowerCase(Locale.ENGLISH)))
-                .map(value -> new Command(null, false, value, "", RequiredType.NONE, null) {})
-                .collect(Collectors.toList());
+                .map(value -> new Command(null, false, value, "", RequiredType.NONE, null) {
+                }).collect(Collectors.toList());
         } else if (args.length > 1) {
             // Additional checks
             Plot plot = player.getCurrentPlot();
@@ -223,7 +229,8 @@ public class Set extends SubCommand {
             }
 
             // components
-            HashSet<String> components = new HashSet<>(Arrays.asList(plot.getManager().getPlotComponents(plot.getId())));
+            HashSet<String> components =
+                new HashSet<>(Arrays.asList(plot.getManager().getPlotComponents(plot.getId())));
             if (components.contains(args[0].toLowerCase())) {
                 return this.component.tab(player, newArgs, space);
             }
