+++ Core/src/main/java/com/plotsquared/core/plot/flag/FlagContainer.java
@@ -146,9 +146,11 @@ public class FlagContainer {
             this.updateSubscribers
                 .forEach(subscriber -> subscriber.handle(flag, plotFlagUpdateType));
         } catch (IllegalStateException e) {
-            PlotSquared.log(String.format("Flag '%s' (class: '%s') could not be added to the container"
-                + " because the flag name exceeded the allowed limit of 64 characters."
-                + " Please tell the developer of that flag to fix this.", flag.getName(), flag.getClass().getName()));
+            PlotSquared.log(String.format(
+                "Flag '%s' (class: '%s') could not be added to the container"
+                    + " because the flag name exceeded the allowed limit of 64 characters."
+                    + " Please tell the developer of that flag to fix this.", flag.getName(),
+                flag.getClass().getName()));
             e.printStackTrace();
         }
     }
