+++ Core/src/main/java/com/plotsquared/core/util/WorldUtil.java
@@ -149,7 +149,7 @@ public abstract class WorldUtil {
                         int trx = top.getX() >> 9;
                         int trz = top.getZ() >> 9;
                         Set<BlockVector2> files =
-                            ChunkManager.manager.getChunkChunks(bot.getWorld());
+                            RegionManager.manager.getChunkChunks(bot.getWorld());
                         for (BlockVector2 mca : files) {
                             if (mca.getX() >= brx && mca.getX() <= trx && mca.getZ() >= brz
                                 && mca.getZ() <= trz) {
